requirements.txt
absl-py==0.6.1
astor==0.7.1
cycler==0.10.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.1
numpy==1.15.3
opencv-python==3.4.3.18
Pillow==5.3.0
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
six==1.11.0
tensorboard==1.12.0
tensorboardX==1.4
tensorflow-gpu==1.11.0
termcolor==1.1.0
torch==0.4.1
torchvision==0.2.1
Werkzeug==0.14.1
